+++ testapps/GHIssues/src/main/java/ghissues/Gh809.java
@@ -2,7 +2,6 @@ package ghissues;
 
 import static act.controller.Controller.Util.template;
 
-import act.Act;
 import act.controller.annotation.UrlContext;
 import org.osgl.mvc.annotation.GetAction;
 
@@ -11,7 +10,6 @@ public class Gh809 extends BaseController {
 
     @GetAction
     public void index() {
-        System.out.println(!Act.isDev());
         template();
     }